Add files via upload
When there is a large number of transactions among multiple persons, it is difficult to keep a track of the cash flows and how much one owe and to whom. Hence there is a need to simplify the transactions among those persons.
The Splitwise Debt Simplification algorithm is designed to solve such problems. We need to implement that algorithm to find out a simple cash flow network. We have implemented concepts of heap and graph, and used HTML, CSS and JavaScript as tech stack. Node of the graph represents individual and the directed edges represents the cash in or cash out.